% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_osESD.R
\name{auto_osESD}
\alias{auto_osESD}
\title{Title}
\usage{
auto_osESD(
  data,
  labeled = FALSE,
  parameter_learning_length = 0.2,
  parameters = parse_args(OptionParser(option_list = list())),
  weights = c(0, 0, 1, 0),
  min_max_switch = FALSE
)
}
\arguments{
\item{data, }{dataset used for anomaly detection. x label 'value', y label 'anomaly'}

\item{labeled, }{TRUE for supervised learning, FALSE for unsupervised learning}

\item{parameter_learning_length, }{length for initiating tuning, default 20%}

\item{parameters, }{parameters, either list or parse arguments}

\item{weights, }{list of weights for scoring, (precision, recall, f1-score, time)}

\item{min_max_switch, }{TRUE for only minimum and maximum values in parameters, FALSE for full parameters}
}
\value{
indices of anomalies
}
\description{
Title
}
