% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osESD.R
\name{osESD}
\alias{osESD}
\title{Title}
\usage{
osESD(
  data,
  size = 100,
  dwin = 5,
  rwin = 5,
  maxr = 10,
  alpha = 0.01,
  condition = TRUE,
  visualize = FALSE
)
}
\arguments{
\item{visualize}{}
}
\description{
Title
}
