% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PRFTcalculator}
\alias{PRFTcalculator}
\title{Title}
\usage{
PRFTcalculator(Reals, Preds, t1, t2)
}
\arguments{
\item{Reals, }{indices of real anomalies}

\item{Preds, }{indices of predicted anomalies}

\item{t1, }{starting time}

\item{t2, }{ending time}
}
\value{
precision, recall, f1score, and run time of run method
}
\description{
Title
}
