% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_osESD.R
\name{auto_osESD}
\alias{auto_osESD}
\title{Title}
\usage{
auto_osESD(
  data,
  labeled = FALSE,
  parameter_learning_length = 0.2,
  parameters = parse_args(OptionParser(option_list = list())),
  weights = c(0, 0, 1, 0),
  min_max_switch = FALSE
)
}
\arguments{
\item{min_max_switch}{}
}
\description{
Title
}
